Rename Wizard component folder
to match Component name
